+++ frontend/mp4write.h
@@ -75,7 +75,7 @@ uint32_t mp4_frame_count(void);
 uint64_t mp4_sample_count(void);
 uint32_t mp4_max_bitrate(void);
 uint32_t mp4_avg_bitrate(void);
-uint32_t mp4_max_frame_size(void);
+uint16_t mp4_max_frame_size(void);
 
 /* Check image header magic bytes (PNG, JPEG, GIF), used to validate
    --cover-art data before it's embedded as an MP4 covr atom. */
